Improve Gemini mocking (#255)
feat(gemini): Improve request matchers for Gemini
feat(mokksy): Add predicate and assert descriptions
fix(mokksy): Remove matched stub immediately to reduce the chance of
race conditions
docs(gemini): Update Gemini mock documentation

+++ ai-mocks-gemini/src/commonMain/kotlin/me/kpavlov/aimocks/gemini/Model.kt
@@ -68,7 +68,7 @@ public data class Part(
 public data class GenerationConfig(
     val temperature: Double? = null,
     val topP: Double? = null,
-    val topK: Int? = null,
+    val topK: Float? = null,
     val candidateCount: Int? = null,
     val maxOutputTokens: Int? = null,
     val stopSequences: List<String>? = null,
